Extracts a FeatureFlagClient interface and abstract base class This will enable a different config resolver to be used in place of the client when needed
